feats: 完成个人信息管理和技能存储

+++ taskcraft-ai-task-infrastructure/src/main/java/tech/neychoup/infrastructure/dao/po/TaskPO.java
@@ -35,7 +35,7 @@ public class TaskPO {
     /**
      * 任务难度
      */
-    private Long difficulty;
+    private Integer difficulty;
 
     /**
      * 代币奖励
